% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_covariate_columns}
\alias{clean_covariate_columns}
\title{Clean covariate Columns
.This converts categorical data to 1-hot encoded columns, and retains numeric column}
\usage{
clean_covariate_columns(df, cols, drop_first = TRUE)
}
\arguments{
\item{df}{Data.frame of covariates for this dataset}

\item{cols}{columns to clean; columns not mentioned here are dropped}

\item{drop_first}{default TRUE; drop first level of encoding for 1-hot}
}
\value{
dataframe of recoded covariates
}
\description{
Clean covariate Columns
.This converts categorical data to 1-hot encoded columns, and retains numeric column
}
\examples{
df = data.frame(x=c("A", "B","C", "A", "B"))
clean_covariate_columns(df,cols="x")
clean_covariate_columns(df,cols="x", drop_first=FALSE)

}
