% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLORAL.R
\name{FLORAL}
\alias{FLORAL}
\title{Fit Log-ratio lasso regression for compositional covariates}
\usage{
FLORAL(
  x,
  y,
  ncov = 0,
  family = "gaussian",
  longitudinal = FALSE,
  id = NULL,
  tobs = NULL,
  failcode = NULL,
  corstr = "exchangeable",
  scalefix = FALSE,
  scalevalue = 1,
  pseudo = 1,
  length.lambda = 100,
  lambda.min.ratio = NULL,
  ncov.lambda.weight = 0,
  a = 1,
  mu = 1,
  pfilter = 0,
  maxiter = 100,
  ncv = 5,
  ncore = 1,
  intercept = FALSE,
  foldid = NULL,
  step2 = TRUE,
  progress = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{Feature matrix, where rows specify subjects and columns specify features. The first \code{ncov} columns should be patient characteristics and the rest columns are microbiome absolute counts corresponding to various taxa. If \code{x} contains longitudinal data, the rows must be sorted in the same order of the subject IDs used in \code{y}.}

\item{y}{Outcome. For a continuous or binary outcome, \code{y} is a vector. For survival outcome, \code{y} is a \code{Surv} object.}

\item{ncov}{An integer indicating the number of first \code{ncov} columns in \code{x} that will not be subject to the zero-sum constraint.}

\item{family}{Available options are \code{gaussian}, \code{binomial}, \code{poisson}, \code{cox}, \code{finegray}. \code{poisson} fits a log-ratio lasso regression with a log ("exponential") link for count outcomes.}

\item{longitudinal}{\code{TRUE} or \code{FALSE}, indicating whether longitudinal data matrix is specified for input \code{x}. (\code{Longitudinal=TRUE} and \code{family="cox"} or \code{"finegray"} will fit a time-dependent covariate model. \code{Longitudinal=TRUE} and \code{family="gaussian"}, \code{"binomial"} or \code{"poisson"} will fit a GEE model.)}

\item{id}{If \code{longitudinal} is \code{TRUE}, \code{id} specifies subject IDs corresponding to the rows of input \code{x}.}

\item{tobs}{If \code{longitudinal} is \code{TRUE}, \code{tobs} specifies time points corresponding to the rows of input \code{x}.}

\item{failcode}{If \code{family = finegray}, \code{failcode} specifies the failure type of interest. This must be a positive integer.}

\item{corstr}{If a GEE model is specified, then \code{corstr} is the corresponding working correlation structure. Options are \code{independence}, \code{exchangeable}, \code{AR-1} and \code{unstructured}.}

\item{scalefix}{\code{TRUE} or \code{FALSE}, indicating whether the scale parameter is estimated or fixed if a GEE model is specified.}

\item{scalevalue}{Specify the scale parameter if \code{scalefix=TRUE}.}

\item{pseudo}{Pseudo count to be added to \code{x} before taking log-transformation. If unspecified, then the log-transformation will not be performed.}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{lambda.min.ratio}{Ratio between the minimum and maximum choice of lambda. Default is \code{NULL}, where the ratio is chosen as 1e-2.}

\item{ncov.lambda.weight}{Weight of the penalty lambda applied to the first \code{ncov} covariates. Default is 0 such that the first \code{ncov} covariates are not penalized.}

\item{a}{A scalar between 0 and 1: \code{a} is the weight for lasso penalty while \code{1-a} is the weight for ridge penalty.}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{pfilter}{A pre-specified threshold to force coefficients with absolute values less than pfilter times the maximum value of absolute coefficient as zeros in the GEE model. Default is zero, such that all coefficients will be reported.}

\item{maxiter}{Number of iterations needed for the outer loop of the augmented Lagrangian algorithm.}

\item{ncv}{Folds of cross-validation. Use \code{NULL} if cross-validation is not wanted.}

\item{ncore}{Number of cores for parallel computing for cross-validation. Default is 1.}

\item{intercept}{\code{TRUE} or \code{FALSE}, indicating whether an intercept should be estimated.}

\item{foldid}{A vector of fold indicator. Default is \code{NULL}.}

\item{step2}{\code{TRUE} or \code{FALSE}, indicating whether a second-stage feature selection for specific ratios should be performed for the features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}

\item{progress}{\code{TRUE} or \code{FALSE}, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{\code{TRUE} or \code{FALSE}, indicating whether returning plots of model fitting.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and others. Details can be found in \code{README.md}.
}
\description{
Conduct log-ratio lasso regression for continuous, binary and survival outcomes.
}
\examples{

set.seed(23420)

# Continuous outcome
dat <- simu(n=50,p=30,model="linear")
fit <- FLORAL(dat$xcount,dat$y,family="gaussian",ncv=2,progress=FALSE,step2=TRUE)

# Binary outcome
# dat <- simu(n=50,p=30,model="binomial")
# fit <- FLORAL(dat$xcount,dat$y,family="binomial",progress=FALSE,step2=TRUE)

# Count outcome (log/exponential link)
# dat <- simu(n=50,p=30,model="poisson")
# fit <- FLORAL(dat$xcount,dat$y,family="poisson",progress=FALSE,step2=TRUE)

# Survival outcome
# dat <- simu(n=50,p=30,model="cox")
# fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d),family="cox",progress=FALSE,step2=TRUE)

# Competing risks outcome
# dat <- simu(n=50,p=30,model="finegray")
# fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d,type="mstate"),failcode=1,
#               family="finegray",progress=FALSE,step2=FALSE)

# Longitudinal continuous outcome
#  dat <- simu(n=50,p=30,model="gee",geetype="gaussian",m=3,corstr="exchangeable",sdvec=rep(1,3))
#  fit <- FLORAL(x=cbind(dat$tvec, dat$xcount),y=dat$y,id=dat$id,family="gaussian",
#                ncov=1,longitudinal = TRUE,corstr = "exchangeable",lambda.min.ratio=1e-3,
#                progress=FALSE,step2=FALSE)

}
\references{
Fei T, Funnell T, Waters N, Raj SS et al. Enhanced Feature Selection for Microbiome Data using FLORAL: Scalable Log-ratio Lasso Regression bioRxiv 2023.05.02.538599.
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
