% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRatioFGLasso.R
\name{LogRatioFGLasso}
\alias{LogRatioFGLasso}
\title{LogRatioFGLasso: Cox log-ratio lasso regression for competing risks outcome}
\usage{
LogRatioFGLasso(
  x,
  y,
  id,
  weight,
  length.lambda = 100,
  lambda.min.ratio = NULL,
  mu = 1,
  ncv = 5,
  foldid = NULL,
  step2 = FALSE,
  progress = TRUE,
  plot = TRUE,
  mcv = "Deviance"
)
}
\arguments{
\item{x}{Covariate data matrix}

\item{y}{A triplet `survival::Surv` object indicating the time period that each microbiome sample represents.}

\item{id}{A covariate of subject IDs.}

\item{weight}{Weights for Fine-Gray model generated by `coxph::finegray()`.}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{lambda.min.ratio}{Ratio between the minimum and maximum choice of lambda. Default is `NULL`, where the ratio is chosen as 1e-2 if n < p and 1e-4 otherwise.}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.}

\item{foldid}{A vector of fold indicator. Default is `NULL`.}

\item{step2}{TRUE or FALSE, indicating whether a stepwise feature selection should be performed for features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}

\item{progress}{TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{TRUE or FALSE, indicating whether returning plots of model fitting.}

\item{mcv}{Metric for cross validation prediction assessment. Default is `Deviance`. An alternative option is `Cindex`.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and many others.
}
\description{
Conduct Cox proportional hazards log-ratio lasso regression
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
