% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRatioLasso.R
\name{LogRatioLasso}
\alias{LogRatioLasso}
\title{LogRatioLasso: Linear log-ratio lasso regression}
\usage{
LogRatioLasso(
  x,
  y,
  length.lambda = 100,
  lambda.min.ratio = NULL,
  mu = 1,
  ncv = 5,
  intercept = FALSE,
  foldid = NULL,
  step2 = FALSE,
  progress = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{Covariate data matrix}

\item{y}{Continuous outcome data vector}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{lambda.min.ratio}{Ratio between the minimum and maximum choice of lambda. Default is `NULL`, where the ratio is chosen as 1e-2 if n < p and 1e-4 otherwise.}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.}

\item{intercept}{TRUE or FALSE, indicating whether an intercept should be estimated.}

\item{foldid}{A vector of fold indicator. Default is `NULL`.}

\item{step2}{TRUE or FALSE, indicating whether a stepwise feature selection should be performed for features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}

\item{progress}{TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{TRUE or FALSE, indicating whether returning plots of model fitting.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and many others.
}
\description{
Conduct linear log-ratio lasso regression
}
\examples{

set.seed(23420)
dat <- simu(n=50,p=100,model="linear")
fit <- LogRatioLasso(dat$x,dat$y,progress=FALSE)

}
\author{
Teng Fei. Email: feit1@mskcc.org
}
