% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRatioLogisticLasso.R
\name{LogRatioLogisticLasso}
\alias{LogRatioLogisticLasso}
\title{LogRatioLogisticLasso: Logistic log-ratio lasso regression}
\usage{
LogRatioLogisticLasso(
  x,
  y,
  length.lambda = 100,
  mu = 1,
  ncv = 5,
  progress = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{Covariate data matrix}

\item{y}{Binary outcome data vector}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.}

\item{progress}{TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{TRUE or FALSE, indicating whether returning plots of model fitting.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and many others.
}
\description{
Conduct logistic log-ratio lasso regression
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
