\name{LogRatioReg-package}
\alias{LogRatioReg-package}
\alias{LogRatioReg}
\docType{package}
\title{
\packageTitle{LogRatioReg}
}
\description{
\packageDescription{LogRatioReg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LogRatioReg}
\packageIndices{LogRatioReg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LogRatioReg}

Maintainer: \packageMaintainer{LogRatioReg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
