% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLORAL.R
\name{FLORAL}
\alias{FLORAL}
\title{FLORAL: Fit Log-ratio lasso regression for compositional covariates}
\usage{
FLORAL(
  x,
  y,
  family = "gaussian",
  longitudinal = FALSE,
  id = NULL,
  tobs = NULL,
  failcode = NULL,
  length.lambda = 100,
  lambda.min.ratio = NULL,
  mu = 1,
  ncv = 5,
  intercept = FALSE,
  foldid = NULL,
  step2 = TRUE,
  progress = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{Count data matrix, where rows specify subjects and columns specify features. If \code{x} contains longitudinal data, the rows must be sorted in the same order of the subject IDs used in \code{y}.}

\item{y}{Outcome. For a continuous or binary outcome, \code{y} is a vector. For survival outcome, \code{y} is a \code{Surv} object.}

\item{family}{Available options are \code{gaussian}, \code{binomial}, \code{cox}, \code{finegray}.}

\item{longitudinal}{TRUE or FALSE, indicating whether longitudinal data matrix is specified for input \code{x}. (Still under development. Please use with caution)}

\item{id}{If \code{longitudinal} is TRUE, \code{id} specifies subject IDs corresponding to the rows of input \code{x}.}

\item{tobs}{If \code{longitudinal} is TRUE, \code{tobs} specifies time points corresponding to the rows of input \code{x}.}

\item{failcode}{If \code{family = finegray}, \code{failcode} specifies the failure type of interest. This must be a positive integer.}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{lambda.min.ratio}{Ratio between the minimum and maximum choice of lambda. Default is \code{NULL}, where the ratio is chosen as 1e-2.}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use \code{NULL} if cross-validation is not wanted.}

\item{intercept}{TRUE or FALSE, indicating whether an intercept should be estimated.}

\item{foldid}{A vector of fold indicator. Default is \code{NULL}.}

\item{step2}{TRUE or FALSE, indicating whether a second-stage feature selection for specific ratios should be performed for the features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}

\item{progress}{TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{TRUE or FALSE, indicating whether returning plots of model fitting.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and others. Details can be found in \code{README.md}.
}
\description{
Conduct log-ratio lasso regression for continuous, binary and survival outcomes.
}
\examples{

set.seed(23420)

# Continuous outcome
dat <- simu(n=50,p=30,model="linear")
fit <- FLORAL(dat$xcount,dat$y,family="gaussian",progress=FALSE,step2=TRUE)

# Binary outcome
dat <- simu(n=50,p=30,model="binomial")
fit <- FLORAL(dat$xcount,dat$y,family="binomial",progress=FALSE,step2=TRUE)

# Survival outcome
dat <- simu(n=50,p=30,model="cox")
fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d),family="cox",progress=FALSE,step2=TRUE)

# Competing risks outcome
dat <- simu(n=50,p=30,model="finegray")
fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d,type="mstate"),failcode=1,
              family="finegray",progress=FALSE,step2=FALSE)

}
\author{
Teng Fei. Email: feit1@mskcc.org
}
