% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLORAL.R
\name{FLORAL}
\alias{FLORAL}
\title{FLORAL: Fit Log-ratio lasso regression for compositional covariates}
\usage{
FLORAL(
  x,
  y,
  family = "gaussian",
  longitudinal = FALSE,
  id = NULL,
  tobs = NULL,
  failcode = NULL,
  length.lambda = 100,
  lambda.min.ratio = NULL,
  mu = 1,
  ncv = 5,
  intercept = FALSE,
  foldid = NULL,
  step2 = TRUE,
  progress = TRUE,
  plot = TRUE,
  table = FALSE
)
}
\arguments{
\item{x}{Count data matrix, where rows specify subjects and columns specify features. If `x` contains longitudinal data, the rows must be sorted in the same order of the subject IDs used in `y`.}

\item{y}{Outcome. For a continuous or binary outcome, `y` is a vector. For survival outcome, `y` is a `Surv` object.}

\item{family}{Available options are `gaussian`, `binomial`, `cox`, `finegray`. `finegray` is still under development. Please use with caution.}

\item{longitudinal}{TRUE or FALSE, indicating whether longitudinal data matrix is specified for input `x`. (Still under development. Please use with caution)}

\item{id}{If `longitudinal` is TRUE, `id` specifies subject IDs corresponding to the rows of input `x`.}

\item{tobs}{If `longitudinal` is TRUE, `tobs` specifies time points corresponding to the rows of input `x`.}

\item{failcode}{If `family = finegray`, `failcode` specifies the failure type of interest. This must be a positive integer.}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{lambda.min.ratio}{Ratio between the minimum and maximum choice of lambda. Default is `NULL`, where the ratio is chosen as 1e-2 if n < p and 1e-4 otherwise.}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.}

\item{intercept}{TRUE or FALSE, indicating whether an intercept should be estimated.}

\item{foldid}{A vector of fold indicator. Default is `NULL`.}

\item{step2}{TRUE or FALSE, indicating whether a second-stage feature selection for specific ratios should be performed for the features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}

\item{progress}{TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{TRUE or FALSE, indicating whether returning plots of model fitting.}

\item{table}{TRUE or FALSE, indicating whether returning gtsummary tables for the 2-stage stepwise model results.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and many others.
}
\description{
Conduct log-ratio lasso regression for continuous, binary and survival outcomes.
}
\examples{

set.seed(23420)

# Continuous outcome
dat <- simu(n=50,p=100,model="linear")
fit <- FLORAL(dat$xcount,dat$y,family="gaussian",progress=FALSE,step2=TRUE)

# Binary outcome
dat <- simu(n=50,p=100,model="binomial")
fit <- FLORAL(dat$xcount,dat$y,family="binomial",progress=FALSE,step2=TRUE)

# Survival outcome
dat <- simu(n=50,p=100,model="cox")
fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d),family="cox",progress=FALSE,step2=TRUE)

}
\author{
Teng Fei. Email: feit1@mskcc.org
}
