% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRatioCoxLasso.R
\name{LogRatioCoxLasso}
\alias{LogRatioCoxLasso}
\title{LogRatioCoxLasso: Cox log-ratio lasso regression}
\usage{
LogRatioCoxLasso(x, y, length.lambda = 100, mu = 1, ncv = 5)
}
\arguments{
\item{x}{Covariate data matrix}

\item{y}{A `survival::Surv` object}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and many others.
}
\description{
Conduct Cox proportional hazards log-ratio lasso regression
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
