% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRatioLasso.R
\name{LogRatioLasso}
\alias{LogRatioLasso}
\title{LogRatioLasso: Linear log-ratio lasso regression}
\usage{
LogRatioLasso(
  x,
  y,
  length.lambda = 100,
  mu = 1,
  ncv = 5,
  intercept = FALSE,
  foldid = NULL,
  progress = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{Covariate data matrix}

\item{y}{Continuous outcome data vector}

\item{length.lambda}{Number of penalty parameters used in the path}

\item{mu}{Value of penalty for the augmented Lagrangian}

\item{ncv}{Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.}

\item{intercept}{TRUE or FALSE, indicating whether an intercept should be estimated.}

\item{foldid}{A vector of fold indicator. Default is `NULL`.}

\item{progress}{TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.}

\item{plot}{TRUE or FALSE, indicating whether returning plots of model fitting.}
}
\value{
A list with path-specific estimates (beta), path (lambda), and many others.
}
\description{
Conduct linear log-ratio lasso regression
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
